+++ examples/.experimental/next-app-dir/src/app/posts/_data.ts
@@ -4,7 +4,7 @@ import { revalidatePath } from 'next/cache';
 import { notFound, redirect, RedirectType } from 'next/navigation';
 import { z } from 'zod';
 import { addPostSchema, type Post } from './_data.schema';
-import { nextProc } from './_trpc';
+import { nextProc } from './_utils/trpc';
 
 const posts: Post[] = [
   {
